:truck: Rename MyTravelAuthorizationLayout to MyTravelAuthorizationWizardLayout.
Why? To make it clear that it's a wizard layout, and not a regular layout.
